Update timelinejs.json
not sure why the original 'item.format' wasn't working, but it does work properly when we change items to two separate lines. This is a trick if you needed to do subsets of data and you want it to be human-readable - see how we create items as a variable on line 7 and then reuse that variable on line 8 to narrow it down.